+++ libft/src/ft_dot.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-double	dot(t_vect v1, t_vect v2)
+float	dot(t_vect v1, t_vect v2)
 {
-	double prod_scal;
+	float prod_scal;
 
 	prod_scal = (v1.x * v2.x) + (v1.y * v2.y) + (v1.z * v2.z);
 	return (prod_scal);
